+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Remove unused `get_body_message_by_title_from_mail` method
 * Remove unused `get_current_date` method
+* Remove unused `archive_inbox` method
 
 ## 0.2.0 (2021-01-19)