+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Remove unused `get_current_date` method
 * Remove unused `archive_inbox` method
 * Remove unused `get_body_by_subject_email` method
+* Remove unused `get_unread_mails_with_tags` method
 
 ## 0.2.0 (2021-01-19)