+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Remove unused `archive_inbox` method
 * Remove unused `get_body_by_subject_email` method
 * Remove unused `get_unread_mails_with_tags` method
+* Remove unused `get_labels` method
 
 ## 0.2.0 (2021-01-19)