+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Remove unused `mark_all_unread` method
 * Remove unused `delete_from_sender` method
 * Remove unused `mail_inbox_count` method
+* Remove unused `mail_in_label_with_date` method
 
 ## 0.2.0 (2021-01-19)