+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Remove unused `delete_from_sender` method
 * Remove unused `mail_inbox_count` method
 * Remove unused `mail_in_label_with_date` method
+* Remove unused `reply_mail` method
 
 ## 0.2.0 (2021-01-19)