+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Remove unused `mail_in_label_with_date` method
 * Remove unused `reply_mail` method
 * Remove unused `delete_all_message_contains` method
+* Remove unused `delete_message_with_portal_address` method
 
 ## 0.2.0 (2021-01-19)