+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Remove unused `reply_mail` method
 * Remove unused `delete_all_message_contains` method
 * Remove unused `delete_message_with_portal_address` method
+* Remove unused `check_unread_messages_for_message` method
 
 ## 0.2.0 (2021-01-19)