+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Remove unused `check_unread_messages_for_message` method
 * Remove unused `get_unread_messages` method
 * Remove unused `wait_until_unread_message` method
+* Remove unused `refresh` method
 
 ## 0.2.0 (2021-01-19)